# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: List tasks with filters
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed, archived]
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: isFavorite
          in: query
          required: false
          schema:
            type: ["boolean", "null"]
        - name: isArchived
          in: query
          required: false
          schema:
            type: ["boolean", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: filter
          in: query
          required: false
          description: "Shorthand filter: today, tomorrow, week, month, overdue, upcoming"
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags: [tasks]
      operationId: createTask
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /tasks/{id}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: Get a single task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /tasks/{id}/duplicate:
    post:
      tags: [tasks]
      operationId: duplicateTask
      summary: Duplicate a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Categories ────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Task schemas ────────────────────────────────────────────────────────
    SubtaskItem:
      type: object
      required: [id, title, completed]
      properties:
        id:
          type: string
        title:
          type: string
        completed:
          type: boolean
    ChecklistItem:
      type: object
      required: [id, text, completed]
      properties:
        id:
          type: string
        text:
          type: string
        completed:
          type: boolean
    Task:
      type: object
      required:
        - id
        - userId
        - title
        - status
        - priority
        - isFavorite
        - isArchived
        - isPinned
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, archived]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: ["string", "null"]
          format: date
        dueTime:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        actualMinutes:
          type: ["integer", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [easy, medium, hard, null]
        energyLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        color:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/SubtaskItem"
        checklistItems:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        isFavorite:
          type: boolean
        isArchived:
          type: boolean
        isPinned:
          type: boolean
        recurringRule:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, archived]
          default: pending
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        categoryId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: ["string", "null"]
          format: date
        dueTime:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [easy, medium, hard, null]
        energyLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        color:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/SubtaskItem"
        checklistItems:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        isFavorite:
          type: boolean
          default: false
        isPinned:
          type: boolean
          default: false
        recurringRule:
          type: ["string", "null"]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, archived]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        categoryId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: ["string", "null"]
          format: date
        dueTime:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        actualMinutes:
          type: ["integer", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [easy, medium, hard, null]
        energyLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        color:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/SubtaskItem"
        checklistItems:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        isFavorite:
          type: boolean
        isArchived:
          type: boolean
        isPinned:
          type: boolean
        recurringRule:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time

    # ── Category schemas ────────────────────────────────────────────────────
...