# lib/api-zod/src/generated/types/checklistItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChecklistItem {
  id: string;
  text: string;
  completed: boolean;
}